+++ unit_converter/utils.py
@@ -1,6 +1,18 @@
 import calendar
 import re
 
+from unit_converter.data_models import UnitData
+
+def get_unit_group(data, name=None) -> str:
+    """Gets unit group"""
+    unit_group: str = input("Unit group: ").strip().lower()
+    if not unit_group:
+        raise ValueError("Unit group cannot be empty!")
+    if name != "manage_type":
+        if unit_group not in data.units:
+            raise KeyError(f"{unit_group} is not a valid group!")
+    return unit_group
+
 
 def resolve_aliases(data, unit_group, unit_type):
     """Checks user's input for any match with unit type or unit's aliases"""